requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements_dev.in
#
aiohttp==3.8.4
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
build==0.10.0
    # via pip-tools
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via pip-tools
cmake==3.26.3
    # via triton
filelock==3.12.0
    # via
    #   torch
    #   triton
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.5.0
    # via pytorch-lightning
idna==3.4
    # via
    #   requests
    #   yarl
jinja2==3.1.2
    # via torch
lightning-utilities==0.8.0
    # via pytorch-lightning
lit==16.0.5.post0
    # via triton
markupsafe==2.1.3
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
networkx==3.1
    # via torch
numpy==1.24.3
    # via
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
packaging==23.1
    # via
    #   build
    #   lightning-utilities
    #   pytorch-lightning
    #   torchmetrics
pillow==9.5.0
    # via torchvision
pip-tools==6.13.0
    # via -r requirements_dev.in
pyproject-hooks==1.0.0
    # via build
pytorch-lightning==2.0.2
    # via -r requirements.in
pyyaml==6.0
    # via pytorch-lightning
requests==2.31.0
    # via
    #   fsspec
    #   torchvision
sympy==1.12
    # via torch
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
torch==2.0.1
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
    #   triton
torchmetrics==0.11.4
    # via pytorch-lightning
torchvision==0.15.2
    # via -r requirements.in
tqdm==4.65.0
    # via pytorch-lightning
triton==2.0.0
    # via torch
typing-extensions==4.6.3
    # via
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
    #   torchmetrics
urllib3==2.0.2
    # via requests
wheel==0.40.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
    #   pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
